requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
boto3==1.38.23
botocore==1.38.23
certifi==2025.4.26
distro==1.9.0
Django==5.2.1
django-cors-headers==4.7.0
django-rest-framework==0.1.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
jmespath==1.0.1
mysqlclient==2.2.7
openai==1.79.0
packaging==25.0
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
